Splitting result in 2 activities
