+++ px-checkout/src/main/java/com/mercadopago/android/px/internal/features/guessing_card/GuessingCardActivityView.java
@@ -101,7 +101,9 @@ public interface GuessingCardActivityView extends MvpView {
     void finishCardFlow(@Nullable PaymentMethod paymentMethod, Token token, Issuer issuer,
         PayerCost payerCost);
 
-    void finishCardStorageFlow(boolean isError, String accessToken);
+    void finishCardStorageFlowWithSuccess();
+
+    void finishCardStorageFlowWithError(String accessToken);
 
     void showProgress();
